fix: update package references after pkg/app→pkg/rpc rename
- Change package name from 'app' to 'rpc' in pkg/rpc/
- Fix references in subcommand/run/tui/event_renderer.go
- Fix references in pkg/cli/rpc.go
- Move ParseSystemPrompt from pkg/cli/helpers.go to subcommand/helpers/
- Add helpers import to pkg/cli/run.go

All tests passing.
