+++ pkg/rpc/rpc_app.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"encoding/json"
