+++ pkg/rpc/rpc_config_handlers.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"fmt"
