+++ pkg/rpc/rpc_format.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"fmt"
