+++ pkg/rpc/rpc_help_handlers.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"fmt"
