+++ pkg/rpc/rpc_message_handlers.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"context"
