+++ pkg/rpc/rpc_server.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"bufio"
