+++ pkg/rpc/rpc_settings.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"encoding/json"
