+++ pkg/rpc/rpc_setup.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"fmt"
