+++ pkg/rpc/rpcapp_smoke_test.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"bufio"
