+++ pkg/rpc/session_writer.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"context"
