+++ pkg/rpc/session_writer_test.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"testing"
