+++ pkg/rpc/traceevent_handler_test.go
@@ -1,4 +1,4 @@
-package app
+package rpc
 
 import (
 	"fmt"
