+++ subcommand/run/tui/event_renderer.go
@@ -186,7 +186,7 @@ func renderSessions(dataJSON []byte) *FormattedEvent {
 // renderSkills renders /skills output.
 func renderSkills(dataJSON []byte) *FormattedEvent {
 	var payload struct {
-		Commands []app.SlashCommand `json:"commands"`
+		Commands []rpc.SlashCommand `json:"commands"`
 	}
 	if err := json.Unmarshal(dataJSON, &payload); err != nil {
 		return fallbackJSON(dataJSON)
@@ -224,8 +224,8 @@ func renderSkills(dataJSON []byte) *FormattedEvent {
 // renderContext renders /context output.
 func renderContext(dataJSON []byte) *FormattedEvent {
 	var payload struct {
-		State  *app.SessionState `json:"state"`
-		Stats  *app.SessionStats `json:"stats"`
+		State  *rpc.SessionState `json:"state"`
+		Stats  *rpc.SessionStats `json:"stats"`
 		Models struct {
 			Models []config.ModelInfo `json:"models"`
 		} `json:"models"`
@@ -314,7 +314,7 @@ func renderContext(dataJSON []byte) *FormattedEvent {
 
 // renderSessionState renders /session output.
 func renderSessionState(dataJSON []byte) *FormattedEvent {
-	var state app.SessionState
+	var state rpc.SessionState
 	if err := json.Unmarshal(dataJSON, &state); err != nil {
 		return fallbackJSON(dataJSON)
 	}
@@ -398,7 +398,7 @@ func renderSessionState(dataJSON []byte) *FormattedEvent {
 // renderMessages renders /messages output.
 func renderMessages(dataJSON []byte) *FormattedEvent {
 	// Try new MessagesResult format first (has total/showing/preview fields)
-	var result app.MessagesResult
+	var result rpc.MessagesResult
 	if err := json.Unmarshal(dataJSON, &result); err == nil && result.Total > 0 {
 		return renderFormattedMessages(result)
 	}
@@ -427,13 +427,13 @@ func renderMessages(dataJSON []byte) *FormattedEvent {
 	}
 
 	// Convert legacy format to MessagesResult
-	legacyResult := app.MessagesResult{
+	legacyResult := rpc.MessagesResult{
 		Total:    len(payload.Messages),
 		Showing:  len(payload.Messages),
-		Messages: make([]app.FormattedMessage, len(payload.Messages)),
+		Messages: make([]rpc.FormattedMessage, len(payload.Messages)),
 	}
 	for i, msg := range payload.Messages {
-		legacyResult.Messages[i] = app.FormattedMessage{
+		legacyResult.Messages[i] = rpc.FormattedMessage{
 			Index:   i,
 			Role:    msg.Role,
 			Preview: msg.Content,
@@ -443,7 +443,7 @@ func renderMessages(dataJSON []byte) *FormattedEvent {
 }
 
 // renderFormattedMessages renders a MessagesResult into a human-readable display.
-func renderFormattedMessages(result app.MessagesResult) *FormattedEvent {
+func renderFormattedMessages(result rpc.MessagesResult) *FormattedEvent {
 	if len(result.Messages) == 0 {
 		return &FormattedEvent{Kind: KindMeta, Text: "no messages"}
 	}
@@ -484,7 +484,7 @@ func renderFormattedMessages(result app.MessagesResult) *FormattedEvent {
 
 // renderModel renders /model output.
 func renderModel(dataJSON []byte) *FormattedEvent {
-	var result app.CycleModelResult
+	var result rpc.CycleModelResult
 	if err := json.Unmarshal(dataJSON, &result); err != nil {
 		// Fallback: try just {model: {id, name}}
 		var payload struct {
@@ -600,7 +600,7 @@ func renderSettings(dataJSON []byte) *FormattedEvent {
 
 // renderSessionStats renders /session stats output.
 func renderSessionStats(dataJSON []byte) *FormattedEvent {
-	var stats app.SessionStats
+	var stats rpc.SessionStats
 	if err := json.Unmarshal(dataJSON, &stats); err != nil {
 		return fallbackJSON(dataJSON)
 	}
@@ -649,7 +649,7 @@ func renderTraceEvents(dataJSON []byte) *FormattedEvent {
 // renderTree renders /tree output.
 func renderTree(dataJSON []byte) *FormattedEvent {
 	var payload struct {
-		Entries []app.TreeEntry `json:"entries"`
+		Entries []rpc.TreeEntry `json:"entries"`
 	}
 	if err := json.Unmarshal(dataJSON, &payload); err != nil {
 		return fallbackJSON(dataJSON)
